Raise an error if a key is not in the map.
Previously, if a key was not in the map, then get_cmd would handle this in its implementation and return None.  But this does not let the caller know whether or not the key is in the map.  By raising a KeyError, the caller will know for sure that the key is not in the map.  This leads to cleaner code for handling keys.