Implement NullBrowser and redo BrowserBuffer's remove.
NullBrowser is a Browser that does nothing.  This is very helpful for
working with Browsers, as a check for whether a Browser is valid or not
is now unecessary.

BrowserBuffer's redo system has changed.  It has become simpler to
maintain.  All remove methods call remove_from_id.  remove_from_id then
makes sure that the table is removable and performs the removals before
calling remove_cleanup.